+++ Documentation/changelog.rst
@@ -11,6 +11,12 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-connector/compare/v2.0.0...HEAD>`_
 -------------------------------------------------------------------------------------------------
 
+Changed
+^^^^^^^
+
+
+* JobRouter REST Client in version 3 is required
+
 `2.0.0 <https://github.com/brotkrueml/typo3-jobrouter-connector/compare/v1.2.0...v2.0.0>`_ - 2023-05-31
 -----------------------------------------------------------------------------------------------------------